Temporary directory fixture (#190)
## Summary

Resolves #173

We cover the builtin fixtures from pytest that are "tmp_path" seen
[here](https://docs.pytest.org/en/7.1.x/builtin.html)

And we allow two more fixture names "temp_path" and "temp_dir".

We do not support "tmpdir" since this is different to "tmp_path" in
pytest and it is not recommended to use it.

## Test Plan

Add tests testing that the fixture gives us the correct object type and
a directory that actually exists.