Use ruff diagnostics (#275)
## Summary

Large change to totally rework our diagnostics.

Frankly we should not care too much about how diagnostics are
constructed, as long as they look nice and emit the important
information.

Ruff already has done lots of work on good diagnostics.

Here we use their API and in turn emit more useful diagnostics.

We will iterate on this in the future, but for now i think this is a
great change.